Add expanded UI components for notifications, battery, and music controls
- Implemented BatteryExpanded.kt for displaying battery status with animations.
- Created ExpandedActions.kt to handle notification actions and intents.
- Developed IncomingCallExpanded.kt for incoming call notifications with action buttons.
- Added IslandExpandedContent.kt to manage the display of various notification types in a pager.
- Introduced MusicExpanded.kt for music playback controls, including play, pause, and repeat functionalities.
- Created NotificationExpanded.kt for displaying general notifications with action buttons and app icons.
