+++ .gitignore
@@ -39,4 +39,5 @@ socials.txt
 
 # Critical README.md
 SMART_ISLAND_REVIEW.md
-FIXES.md
+BUGS_AND_FIXES.md
+SmartIsland_AUDIT_FIXES.md
